requirements.txt
Django==5.0.0
django-tailwind==3.8.0
django-crispy-forms==2.1
crispy-tailwind==1.0.3
Pillow==10.1.0
django-widget-tweaks==1.5.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
google-api-python-client==2.110.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openai==1.6.0
pydub==0.25.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
python-dotenv==1.0.0
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1

# Filtering
django-filter==23.5

# Testing
pytest-django==4.7.0
pytest-cov==4.1.0
coverage==7.3.4

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Security
django-ratelimit==4.1.0
django-cors-headers==4.3.1

# Monitoring
django-health-check==3.18.1
sentry-sdk==1.39.2

# Performance
django-debug-toolbar==4.2.0
